[Connector][WRS-3099] Split Google Sheet connector into multiple modules
