+++ src/connectors/google-sheet/connector.ts
@@ -3,395 +3,19 @@ import {
   Data,
   BidirectionalDataPageItem,
 } from '@chili-publish/studio-connectors';
+import { Converter } from './lib/converter';
+import { RangeHelper } from './lib/range-helper';
+import {
+  CONTEXT_SPREADSHEET_URL_PROPERTY,
+  ITEM_ROW_ID_PROPERTY,
+  RowIdHelper,
+} from './lib/row-id-helper';
 import type {
-  NumberCell,
-  DateCell,
-  BooleanCell,
+  ApiError,
   CellData,
   Row,
-  EmptyRowWithoutFormatting,
   Spreadsheet,
-  ApiError,
-  TypedNumberCell,
-  TypedDateCell,
-  TypedPlainTextCell,
-  TypedBooleanCell,
-} from './types';
-
-class RangeHelper {
-  static buildHeaderRange(sheetName: string | null) {
-    return RangeHelper.buildRange(sheetName, 1, 1);
-  }
-
-  static buildFirstPageRange(sheetName: string | null, limit: number) {
-    // starts from 2 since exclude header row that has index '1'
-    return RangeHelper.buildRange(sheetName, 2, limit > 1 ? 1 + limit : 2);
-  }
-
-  static buildNextPageRange(currentRange: string, limit: number) {
-    const [sheetName, _, lastRow] = RangeHelper.extractFromRange(currentRange);
-    if (Number.isNaN(lastRow)) {
-      throw new Error(`Incorrect format of the cells range "${currentRange}"`);
-    }
-    return RangeHelper.buildRange(sheetName, lastRow + 1, lastRow + limit);
-  }
-
-  static buildPreviousPageRange(currentRange: string, limit: number) {
-    const [sheetName, startRow] = RangeHelper.extractFromRange(currentRange);
-    if (Number.isNaN(startRow)) {
-      throw new Error(`Incorrect format of the cells range "${currentRange}"`);
-    }
-    const prevEndRow = startRow - 1;
-    const prevStartRow = Math.max(2, prevEndRow - limit + 1);
-    return RangeHelper.buildRange(sheetName, prevStartRow, prevEndRow);
-  }
-
-  static buildRowRange(sheetName: string | null, rowNumber: number) {
-    return RangeHelper.buildRange(sheetName, rowNumber, rowNumber);
-  }
-
-  /**
-   * Returns the range string for the next page page given the row number and the page size (limit).
-   * Row numbers are 1-indexed in the sheet; row 1 is the
-   * header, so data rows start at row 2.
-   */
-  static buildRangeForNextPage(
-    sheetName: string | null,
-    rowNumber: number,
-    limit: number
-  ) {
-    const isFirstPage = RangeHelper.isFirstPage(rowNumber, limit);
-
-    if (isFirstPage) {
-      return RangeHelper.buildRange(
-        sheetName,
-        rowNumber + 1,
-        rowNumber + limit
-      );
-    }
-
-    const pageStartRow = rowNumber + 1;
-    const pageEndRow = pageStartRow + limit - 1;
-
-    return RangeHelper.buildRange(sheetName, pageStartRow, pageEndRow);
-  }
-
-  static buildRangeForPreviousPage(
-    sheetName: string | null,
-    rowNumber: number,
-    limit: number
-  ) {
-    const isFirstPage = RangeHelper.isFirstPage(rowNumber, limit);
-
-    if (isFirstPage) {
-      if (rowNumber - 1 < 2) return null; // (row 2 → pos 0)
-      return RangeHelper.buildRange(sheetName, 2, rowNumber - 1);
-    }
-
-    const pageEndRow = rowNumber - 1;
-    const pageStartRow = Math.max(2, pageEndRow - limit + 1);
-
-    return RangeHelper.buildRange(sheetName, pageStartRow, pageEndRow);
-  }
-
-  static getStartRow(range: string): number {
-    const [, startRow] = RangeHelper.extractFromRange(range);
-    return startRow;
-  }
-
-  static getEndRow(range: string): number {
-    const [, , endRow] = RangeHelper.extractFromRange(range);
-    return endRow;
-  }
-
-  /**
-   * Builds a range for `limit` rows starting at startRow. Use when resolving
-   * continuationToken with a possibly changed limit.
-   */
-  static buildRangeFromStartRow(
-    sheetName: string | null,
-    startRow: number,
-    limit: number
-  ): string {
-    return RangeHelper.buildRange(
-      sheetName,
-      startRow,
-      startRow + Math.max(0, limit - 1)
-    );
-  }
-
-  /**
-   * Builds the previous page range when the token encodes the previous page's
-   * range and the request may use a different limit. Returns the last `limit`
-   * rows before the row after tokenEndRow.
-   */
-  static buildPreviousPageRangeFromToken(
-    sheetName: string | null,
-    tokenEndRow: number,
-    limit: number
-  ): string {
-    const prevEndRow = tokenEndRow;
-    const prevStartRow = Math.max(2, prevEndRow - limit + 1);
-    return RangeHelper.buildRange(sheetName, prevStartRow, prevEndRow);
-  }
-
-  private static buildRange(
-    sheetName: string | null,
-    start: number,
-    end: number
-  ) {
-    return sheetName ? `${sheetName}!${start}:${end}` : `${start}:${end}`;
-  }
-
-  static isFirstPage(rowNumber: number, limit: number) {
-    const dataPos = rowNumber - 2; // 0-indexed data position (row 2 → pos 0)
-    const pageIndex = Math.floor(dataPos / limit);
-    return pageIndex === 0;
-  }
-
-  private static extractFromRange(
-    range: string
-  ): [string | null, number, number] {
-    const splitted = range.split('!');
-
-    const sheetName = splitted.length === 1 ? null : splitted[0]; // when we request data without sheetName
-    const cellsQuery = splitted.length === 1 ? splitted[0] : splitted[1];
-
-    const splittedCells = cellsQuery.split(':');
-    return [sheetName, Number(splittedCells[0]), Number(splittedCells[1])];
-  }
-}
-
-const ITEM_ROW_ID_PROPERTY = '__rowId__' as const;
-
-const CONTEXT_SPREADSHEET_URL_PROPERTY = 'spreadsheetURL' as const;
-
-class RowIdHelper {
-  /**
-   * When the URL has no `gid`, the connector targets the first sheet; row ids use `0` for that case.
-   */
-  static canonicalSheetId(sheetId: string | null): string {
-    return sheetId ?? '0';
-  }
-
-  static build(
-    spreadsheetId: string,
-    sheetId: string | null,
-    rowNumber: number
-  ): string {
-    return `${spreadsheetId}_${RowIdHelper.canonicalSheetId(
-      sheetId
-    )}_${rowNumber}`;
-  }
-
-  /**
-   * Parses `{spreadsheetId}_{sheetId}_{rowNumber}`. The spreadsheet id may contain `_`;
-   * sheet id and row are the last two segments (sheet id is numeric).
-   */
-  static parse(id: string): {
-    spreadsheetId: string;
-    sheetId: string;
-    rowNumber: number;
-  } | null {
-    const parts = id.split('_');
-    if (parts.length < 3) {
-      return null;
-    }
-    const rowPart = parts[parts.length - 1]!;
-    const sheetIdPart = parts[parts.length - 2]!;
-    const spreadsheetId = parts.slice(0, -2).join('_');
-    const rowNumber = parseInt(rowPart, 10);
-    if (!Number.isInteger(rowNumber) || rowNumber < 2) {
-      return null;
-    }
-    if (!/^\d+$/.test(sheetIdPart)) {
-      return null;
-    }
-    return { spreadsheetId, sheetId: sheetIdPart, rowNumber };
-  }
-
-  static assignRowIdToItem(
-    item: Data.DataItem,
-    spreadsheetId: string,
-    sheetId: string | null,
-    rowNumber: number
-  ): Data.DataItem {
-    const rowId = RowIdHelper.build(spreadsheetId, sheetId, rowNumber);
-    const rest = { ...(item as Record<string, unknown>) };
-    delete rest[ITEM_ROW_ID_PROPERTY];
-    return {
-      [ITEM_ROW_ID_PROPERTY]: rowId,
-      ...rest,
-    } as Data.DataItem;
-  }
-}
-
-class Converter {
-  static toDataItems(
-    tableHeader: Row<Required<CellData>>,
-    tableBody: Array<Row>
-  ): Array<Data.DataItem> {
-    const tableHeaderValues = tableHeader.values;
-    return (
-      tableBody
-        .map((row) => {
-          const item =
-            this.normalizeRow(row, tableHeaderValues.length)?.values.reduce(
-              (acc, tableCell, colIndex) => {
-                const { type, cell } = Converter.toTypedCell(tableCell);
-
-                switch (type) {
-                  case 'number':
-                    acc[tableHeaderValues[colIndex].formattedValue] =
-                      cell.effectiveValue?.numberValue ?? null;
-                    break;
-                  case 'date':
-                    acc[tableHeaderValues[colIndex].formattedValue] =
-                      this.convertToDate(cell.effectiveValue?.numberValue);
-                    break;
-                  case 'boolean':
-                    acc[tableHeaderValues[colIndex].formattedValue] =
-                      cell.effectiveValue.boolValue;
-                    break;
-                  case 'singleLine':
-                    acc[tableHeaderValues[colIndex].formattedValue] =
-                      cell.formattedValue ?? null;
-                    break;
-                }
-                return acc;
-              },
-              {} as Data.DataItem
-            ) ?? null;
-
-          if (item === null) return null;
-          return item;
-        })
-        // Filter out empty rows
-        .filter((d): d is Data.DataItem => d !== null)
-    );
-  }
-
-  static toDataModelProperties(
-    headerRow: Row<Required<CellData>>,
-    bodyRows: Array<Row>
-  ): Array<Data.DataModelProperty> {
-    if (!bodyRows.length) {
-      throw new Error(
-        'Model can not be generated. To execute the operation your sheet should have the row with data in addition to the header row'
-      );
-    }
-    const normalizedBodyRow = Converter.normalizeRow(
-      bodyRows[0],
-      headerRow.values.length
-    );
-    if (!normalizedBodyRow) {
-      throw new Error(
-        'Model can not be generated. To execute the operation your sheet should have the row with data in addition to the header row'
-      );
-    }
-    const { values } = normalizedBodyRow;
-    return headerRow.values.map((column, idx) => {
-      return {
-        type: Converter.toTypedCell(values[idx]).type,
-        name: column.formattedValue,
-      };
-    });
-  }
-
-  /**
-   * Inspects a Google Sheets cell and determines its semantic type—number, date, boolean, or single line of text—
-   * based on the provided cell's formatting and effective value. This is used to map Google Sheets' flexible cell
-   * data model to strongly typed connector model data.
-   *
-   * @param cell The cell data object returned by the Google Sheets API, possibly containing formatting and value information.
-   */
-  private static toTypedCell(
-    cell: CellData
-  ): TypedNumberCell | TypedDateCell | TypedPlainTextCell | TypedBooleanCell {
-    if (
-      'effectiveFormat' in cell &&
-      'numberFormat' in cell.effectiveFormat &&
-      cell.effectiveFormat.numberFormat.type === 'NUMBER'
-    ) {
-      return {
-        type: 'number',
-        cell: cell as NumberCell,
-      };
-    }
-
-    if (
-      'effectiveFormat' in cell &&
-      'numberFormat' in cell.effectiveFormat &&
-      (cell.effectiveFormat.numberFormat.type === 'DATE' ||
-        cell.effectiveFormat.numberFormat.type === 'DATE_TIME')
-    ) {
-      return {
-        type: 'date',
-        cell: cell as DateCell,
-      };
-    }
-
-    if (
-      'effectiveValue' in cell &&
-      cell.effectiveValue &&
-      'boolValue' in cell.effectiveValue
-    ) {
-      return {
-        type: 'boolean',
-        cell: cell as BooleanCell,
-      };
-    }
-    return {
-      type: 'singleLine',
-      cell: cell,
-    };
-  }
-
-  /**
-   * Depends on whether row contains custom formatting or not Google Sheets API return data in different way
-   * This function is reponsible to handle all such use cases and return Row always in regular form as well as handling empty rows use case
-   *
-   * @param row
-   * @param columnsLength
-   * @returns
-   */
-  private static normalizeRow(
-    row: Row | EmptyRowWithoutFormatting,
-    columnsLength: number
-  ): Row | null {
-    if (!row.values) {
-      return null;
-    }
-    const emptyRow = row.values.every((c) => !c.formattedValue);
-    if (emptyRow) {
-      return null;
-    }
-    if (row.values.length === columnsLength) {
-      return row;
-    }
-    return {
-      values: [
-        ...row.values,
-        ...new Array(columnsLength - row.values.length).fill({}),
-      ],
-    };
-  }
-
-  /**
-   * The number value that Google sheets represent as date refers to serial number of internal date system
-   * This function takes this into account and transofrm to regular date
-   * @param serialNumber Internal date representation
-   */
-  private static convertToDate(serialNumber?: number) {
-    if (!serialNumber) {
-      return null;
-    }
-    // Google Sheets epoch date is December 30, 1899
-    const epoch = new Date(Date.UTC(1899, 11, 30)); // UTC to avoid timezone issues
-    const date = new Date(epoch.getTime() + serialNumber * 24 * 60 * 60 * 1000); // Add days in milliseconds return
-    return date;
-  }
-}
+} from './lib/types';
 
 const FIELDS_MASK = `sheets.properties(sheetId,title),sheets.data.rowData.values(formattedValue,effectiveFormat.numberFormat.type,effectiveValue)`;
 export default class GoogleSheetConnector
